Reduce dataseat to 3000 rows
We have huge datsets at times, often for a day more than 100 tweets were observed which may skew the ratio. Hence we need to agree on a number of tweets per day and select that. Used simple head and concat function to achieve this, the number of topics to choose is to be decided.